Move the email module out of the auth sub module to the app folder
