+++ app/auth/routes.py
@@ -4,10 +4,10 @@ from flask_login import current_user, login_user, logout_user
 
 from app import db
 from app.models import User
+from app.email import send_password_reset_email
 from app.auth import bp
 from app.auth.forms import LoginForm, RegistrationForm, \
     RequestPasswordResetForm, PasswordResetForm
-from app.auth.email import send_password_reset_email
 
 
 @bp.route('/login', methods=['GET', 'POST'])
